requirements.txt
django-model-utils>=2.0
geocoder
django-geoposition
django-annoying
django-anymail[mailgun]
django-celery-beat
bs4
# Additional requirements go here
